# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SaaS API Documentation",
        "description": "API documentation for the SaaS platform with authentication, user management, company management, and plan management.",
        "contact": {
            "name": "API Support",
            "email": "admin@saas.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/plans": {
            "get": {
                "tags": [
                    "Plans"
                ],
                "summary": "List all plans",
                "description": "Get paginated list of plans. Requires super-admin role.",
                "operationId": "listPlans",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plans retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlanListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Insufficient permissions"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Plans"
                ],
                "summary": "Create a new plan",
                "description": "Create a new plan. Requires super-admin role.",
                "operationId": "createPlan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePlanRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plan created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlanResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Insufficient permissions"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/plans/{id}/restore": {
            "patch": {
                "tags": [
                    "Plans"
                ],
                "summary": "Restore plan",
                "description": "Restore a soft deleted plan. Requires super-admin role.",
                "operationId": "restorePlan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Plan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan restored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlanSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Insufficient permissions"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plan not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plan not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreatePlanRequest": {
                "required": [
                    "name",
                    "monthly_price",
                    "user_limit",
                    "features"
                ],
                "properties": {
                    "name": {
                        "description": "Plan name",
                        "type": "string",
                        "example": "Basic Plan"
                    },
                    "monthly_price": {
                        "description": "Monthly price",
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "user_limit": {
                        "description": "Maximum number of users",
                        "type": "integer",
                        "example": 10
                    },
                    "features": {
                        "description": "Plan features",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "feature1",
                            "feature2"
                        ]
                    }
                },
                "type": "object"
            },
...
            "Plan": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Basic Plan"
                    },
                    "monthly_price": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "user_limit": {
                        "type": "integer",
                        "example": 10
                    },
                    "features": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "feature1",
                            "feature2"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "PlanListResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Plan"
                        }
                    },
                    "pagination": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 100
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 7
                            },
                            "from": {
                                "type": "integer",
                                "example": 1
                            },
                            "to": {
                                "type": "integer",
                                "example": 15
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PlanResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Plan created successfully"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Plan"
                    }
                },
                "type": "object"
            },
            "PlanSuccessResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Plan deleted successfully"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the token without the 'Bearer ' prefix",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}